requirements.txt
speechrecognition
pyaudio
pyttsx3
spacy
pyyaml
